third_party/blink/web_tests/external/wpt/dom/nodes/DOMImplementation-createDocument-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] createDocument test: undefined,"f::oo",null,"NAMESPACE_ERR"
  assert_throws_dom: function "function() { document.implementation.createDocument(namespace, qualifiedName, doctype) }" threw object "InvalidCharacterError: Failed to execute 'createDocument' on 'DOMImplementation': The qualified name provided ('f::oo') has an empty local name." that is not a DOMException NAMESPACE_ERR: property "code" is equal to 5, expected 14
[FAIL] createDocument test: "http://example.com/","f:o:o",null,null
  assert_equals: element.localName expected "o:o" but got "o"
[FAIL] createDocument test: "http://example.com/","f::oo",null,null
  Failed to execute 'createDocument' on 'DOMImplementation': The qualified name provided ('f::oo') has an empty local name.
[FAIL] createDocument test: metadata for "http://example.com/","f::oo",null
  Failed to execute 'createDocument' on 'DOMImplementation': The qualified name provided ('f::oo') has an empty local name.
[FAIL] createDocument test: characterSet aliases for "http://example.com/","f::oo",null
  Failed to execute 'createDocument' on 'DOMImplementation': The qualified name provided ('f::oo') has an empty local name.
[FAIL] createDocument test: "http://example.com/","prefix::local",null,null
  Failed to execute 'createDocument' on 'DOMImplementation': The qualified name provided ('prefix::local') has an empty local name.
[FAIL] createDocument test: metadata for "http://example.com/","prefix::local",null
  Failed to execute 'createDocument' on 'DOMImplementation': The qualified name provided ('prefix::local') has an empty local name.
[FAIL] createDocument test: characterSet aliases for "http://example.com/","prefix::local",null
  Failed to execute 'createDocument' on 'DOMImplementation': The qualified name provided ('prefix::local') has an empty local name.
Harness: the test ran to completion.
